+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add page titles to Transfer and Conversion project task pages
 - Add RPA Policy task to Transfer projects
 - Add Form M task to Transfer projects
+- Add Church Supplemental Agreement task to Transfer projects
 
 ### Changed